+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- option `CC1101(lock_spi_device=True)` sets an advisory, exclusive,
+  non-blocking `flock` on `/dev/spidev?.?`
 
 ## [2.2.0] - 2020-12-08
 ### Added